+++ internal/conf/flags/flags.go
@@ -4,7 +4,7 @@ var (
 	FlagConfigFilePath string
 	FlagVerbose        bool
 	FlagSocket         string
-	FlagPort           int
+	FlagPort           int64
 	FlagRootDir        string
 	FlagStateDir       string
 	FlagWorkDir        string
